requirements.txt
# Anthropic SDK for base rate agent
anthropic>=0.39.0

# API clients
httpx>=0.27.0
aiohttp>=3.9.0
websockets>=12.0

# Web3 for Polymarket (CLOB)
py-clob-client>=0.1.0

# Web UI
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0

# Data handling
pandas>=2.2.0
python-dateutil>=2.9.0

# Utilities
pydantic>=2.9.0
python-dotenv>=1.0.0
tenacity>=9.0.0

# Sportsbook odds
requests>=2.32.0
